requirements.txt
django 
djangorestframework
dotenv
mysqlclient
djangorestframework-simplejwt
watchdog
drf-spectacular 
drf-spectacular[sidecar]
